# openapi.yaml
openapi: 3.0.3
info:
  title: LIS Labo-Anapath API
  description: API REST de gestion de laboratoire d'anatomopathologie
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Développement local
  - url: https://api.labo-anapath.com/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Liste des utilisateurs
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageApiResponse'
    post:
      tags: [Users]
      summary: Créer un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /patients:
    get:
      tags: [Patients]
      summary: Liste des patients
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/search'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortDir'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientPageApiResponse'
    post:
      tags: [Patients]
      summary: Créer un patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /patients/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Patients]
      summary: Détail d'un patient
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientApiResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Patients]
      summary: Modifier un patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientApiResponse'
    delete:
      tags: [Patients]
      summary: Supprimer un patient (soft delete)
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── DOCTORS ──────────────────────────────────────────────────────────────────
...
  /suppliers:
    get:
      tags: [Suppliers]
      summary: Liste des fournisseurs
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - $ref: '#/components/parameters/search'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierPageApiResponse'
    post:
      tags: [Suppliers]
      summary: Créer un fournisseur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierApiResponse'
  /suppliers/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Suppliers]
      summary: Détail d'un fournisseur
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierApiResponse'
    put:
      tags: [Suppliers]
      summary: Modifier un fournisseur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupplierApiResponse'
    delete:
      tags: [Suppliers]
      summary: Supprimer un fournisseur
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── MOVEMENTS ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── REUSABLE PARAMETERS ──────────────────────────────────────────────────────
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    search:
      name: search
      in: query
      schema:
        type: string
    sortBy:
      name: sortBy
      in: query
      schema:
        type: string
        default: createdAt
    sortDir:
      name: sortDir
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc

  # ── REUSABLE RESPONSES ───────────────────────────────────────────────────────
  responses:
    EmptySuccess:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmptyApiResponse'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
    Conflict:
      description: Ressource déjà existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
...
    ValidationError:
      description: Données de saisie invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorApiResponse'

  # ── SCHEMAS ──────────────────────────────────────────────────────────────────
  schemas:
...
    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
    EmptyApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Ressource introuvable
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
    ValidationErrorApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Erreurs de validation
        data:
          type: object
          additionalProperties:
            type: string
          example:
            firstName: Le prénom est obligatoire
            email: L'email doit être valide
        timestamp:
          type: string
          format: date-time
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        isActive:
          type: boolean
        branchId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    UserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        phone:
          type: string
        isActive:
          type: boolean
          default: true
    UserApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/UserResponse'
        timestamp: {type: string, format: date-time}
    UserPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/UserResponse'
        timestamp: {type: string, format: date-time}

    # -- Role / Permission ----------------------------------------------------
...
    PatientRequest:
      type: object
      required: [firstName, lastName, gender]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          pattern: '^[+]?[0-9]{6,20}$'
        email:
          type: string
          format: email
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
        branchId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    PatientApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/PatientResponse'
        timestamp: {type: string, format: date-time}
    PatientPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/PatientResponse'
        timestamp: {type: string, format: date-time}

    # -- Doctor / Hospital ----------------------------------------------------
...
    SupplierRequest:
      type: object
      required: [name]
      properties:
        name: {type: string}
        phone: {type: string}
        email: {type: string, format: email}
        address: {type: string}
        category: {type: string}
    SupplierResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        phone: {type: string}
        email: {type: string}
        address: {type: string}
        category: {type: string}
        branchId: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    SupplierApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/SupplierResponse'
        timestamp: {type: string, format: date-time}
    SupplierPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/SupplierResponse'
        timestamp: {type: string, format: date-time}
...